Fix the code issues MSVC /W4 found
Turning on /W4 /WX surfaced 12 warnings in our own code (everything else was
vendored or STL). None were harmless-by-construction, so they are fixed rather
than suppressed.

Real defects:
  - TabView: `Rectangle strip` / `stripClip` were assigned only in switch arms,
    so a path MSVC could not rule out read them uninitialized (C4701). Now
    value-initialized at declaration.
  - Reliable: TransportStats::rttMs is f32 but Connection tracks rtt in f64;
    the assignment narrowed silently. Explicit now.
  - ShaderSystemHost / UISubsystemTests: size_t returned or stored as u32
    without a cast.

Readability, and the warning was right:
  - ParticleEffectPageImpl: the module-inspector `else if (auto* m = Cast<T>())`
    chain reused `m` in 20 nested branches. Each is now named for its type
    (posInit, velInit, gravityMod, collisionMod, ...), which reads better than
    twenty identical `m`s. Re-wrapped with clang-format afterwards.
  - CookServiceImpl: an inner `drained` shadowed an outer one; the inner is the
    queue being drained, so it is `pending`.
  - GUISandbox: an inner `row` (a table row's values) shadowed an outer `row`
    (a LinearLayout of buttons); the inner is `cells`.
  - SelectionModel::FirstSelected used a range-for whose body always returns, so
    the loop increment was provably unreachable (C4702). Iterator form instead.
  - InspectorViewImpl: a compile-time flag was folded into a runtime && , which
    cl 19.44 flags as C4127 (19.51 does not). `if constexpr` on the flag says
    what is meant.

Verified on cl 19.51, cl 19.44 and clang, all with warnings-as-errors.

+++ Code/Draconic/Engine/Engine.UI.Tests/UISubsystemTests.cpp
@@ -612,7 +612,7 @@ TEST_CASE("ui.subsystem: preview roots live in the context but never on the scre
     CHECK(Cast<ViewGroup>(preview.Get())->FindByName(u8"pv") != nullptr);
     // ...which is NOT parented to the screen root (RenderOverlay draws only the screen
     // root, so a preview can never leak into game targets)...
-    const u32 screenChildren = ui->ScreenRoot()->ChildCount();
+    const u32 screenChildren = static_cast<u32>(ui->ScreenRoot()->ChildCount());
     CHECK(Cast<ViewGroup>(ui->ScreenRoot())->FindByName(u8"pv") == nullptr);
     // ...and frames tick without disturbing the screen tier.
     ctx.BeginFrame(1.0f / 60.0f);
